Chore: put example in one file to show it in pubdev example tab normally.
