+++ packages/elementary/example/android/app/src/main/kotlin/com/example/counter/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.counter
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()
