--- packages/elementary/example/android/app/src/main/kotlin/com/example/mwwmcounter/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.mwwmcounter.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
